+++ src/deveeldb/Deveel.Data.Transactions/Transaction.cs
@@ -48,7 +48,7 @@ namespace Deveel.Data.Transactions {
 		private bool autoCommit;
 		private string parameterStyle;
 
-		internal Transaction(Database database, long commitId, TransactionIsolation isolation, IEnumerable<TableSource> committedTables, IEnumerable<IIndexSet> indexSets) {
+		internal Transaction(Database database, int commitId, TransactionIsolation isolation, IEnumerable<TableSource> committedTables, IEnumerable<IIndexSet> indexSets) {
 			CommitId = commitId;
 			Database = database;
 			Isolation = isolation;
@@ -72,7 +72,7 @@ namespace Deveel.Data.Transactions {
 			ignoreCase = database.DatabaseContext.IgnoreIdentifiersCase();
 		}
 
-		internal Transaction(Database database, long commitId, TransactionIsolation isolation)
+		internal Transaction(Database database, int commitId, TransactionIsolation isolation)
 			: this(database, commitId, isolation, new TableSource[0], new IIndexSet[0]) {
 		}
 
@@ -93,7 +93,7 @@ namespace Deveel.Data.Transactions {
 			IntTableInfo[8] = SystemSchema.PrivilegesTableInfo;
 		}
 
-		public long CommitId { get; private set; }
+		public int CommitId { get; private set; }
 
 		object ILockable.RefId {
 			get { return CommitId; }
